Updates and changes:
Main
- Added labels indicating the start and end of every operation.
- Fixed some issues regarding the HoursWorked class, gross wage, and net
wage.

Employee Class
- Added '₱' character to indicate Philippine currency for basic salary
up to hourly rate.
- Removed public double getHourlyRate(); and public int
getEmployeeNo();.

HoursWorked Class
- Added private string, int, and double variables to represent
employeeNo, time, day, month, and year to be used in the calculation.
- Added public HoursWorked(int employeeNo) constructor method to access
and gain access to the parameter "employeeNo" user input in Main.
- Added "this.attribute" to represent the value of private string, int,
and double variables inside the constructor method.
- Added switch case statements to represent "employeeNo" and
"this.attribute" to represent the value for each day in the month.
- Added "this.attribute" to add value to the year, months, variable.
- Added "public void display()" to print the calculation of worked hours
and compare it with required hours.